+++ src/components/CategoryCard.tsx
@@ -1,5 +1,5 @@
-import { Link } from "react-router-dom";
 import styled from "styled-components";
+import { Link } from "react-router-dom";
 
 const Card = styled(Link)<{ nth: number }>`
   background-color: rgba(255, 255, 255, 0.5);
